Migrate to GitHub Actions
* Fix test_stay_connected when stdin is not a tty

That happens on GitHub Actions runners. When stdin is a tty,
LogcatColor.start() runs logcat first, so there is one more adb command
compared to the case when stdin is not a tty. Use an input file for test
consistency on varying environments.

See: https://github.com/yan12125/logcat-color3/issues/2